+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 12.5
 - JS Prevent submitting form if no checkboxes are checked in warehouse.js & Inputs.php
 - JS add bottomButtonClick() Bottom.php buttons click related events in warehouse.js & Bottom.php
 - JS deprecate switchMenu() function: use `<details><summary>` instead in warehouse.js, General_Info.inc.php, Search.fnc.php & Widgets.php
+- JS ajaxPrepare(): handle #ajax_update_body, #menu_user_session & #warehouse_user_session in warehouse.js, Side.php & Warehouse.php
 
 Changes in 12.4.3
 -----------------